% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAA-package.R
\docType{package}
\name{BAA-package}
\alias{BAA}
\alias{BAA-package}
\title{BAA: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
