% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coeff_comp}
\alias{coeff_comp}
\title{Compare linear regression coefficients}
\usage{
coeff_comp(
  model1,
  model2,
  predictor,
  predictor2 = NULL,
  nested = TRUE,
  ci = TRUE,
  pval = TRUE
)
}
\arguments{
\item{model1}{linear regression model object; reduced model if nested}

\item{model2}{linear regression model object; full model if nested}

\item{predictor}{character string name of coefficient for which to compute difference}

\item{predictor2}{if in model2 the coefficient of interest is named something else; NULL by default}

\item{nested}{TRUE by default}

\item{ci}{outputs 95\% confidence interval if TRUE}

\item{pval}{outputs p-value if TRUE}
}
\value{
Returns a vector, which includes difference of coefficients, standard error, and, optionally, 95\%CI, p-value.
}
\description{
Allows to empirically compare coefficients for the same predictor in a linear regression model
either before and after adjustment for covariates (nested models, not independent)
or for different samples (can use the same formula, models assumed independent).
}
\seealso{
Statistical Methods for Comparing Regression Coefficients Between Models
  Author(s): Clifford C. Clogg, Eva Petkova and Adamantios Haritou
  Source: American Journal of Sociology, Vol. 100, No. 5 (Mar., 1995), pp. 1261-1293
  Published by: The University of Chicago PressStable URL: https://www.jstor.org/stable/2782277
}
