% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{csv.to.pub}
\alias{csv.to.pub}
\title{Make csv into publication-ready excel table}
\usage{
csv.to.pub(file.path, file.name)
}
\arguments{
\item{file.path}{path of folder in which file of interest is located}

\item{file.name}{name of csv file (including ".csv")}
}
\value{
Returns a dataframe.
}
\description{
Make csv into publication-ready excel table
}
