% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make.sheet.name}
\alias{make.sheet.name}
\title{Translate the name of the model to something publication-readable}
\usage{
make.sheet.name(file.name)
}
\arguments{
\item{file.name}{name of csv file (including ".csv")}
}
\value{
Returns a string
}
\description{
Make sheet names for the final excel file
}
