% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_margins}
\alias{make_margins}
\title{Produces and plots predictive margins}
\usage{
make_margins(
  object,
  newdata,
  x_axis = "",
  y_axis = "predicted",
  grouping = "",
  x_range = seq(-2, 6, 1),
  which_smooth = "both",
  point_shape = 16,
  bandwidth = 0.5,
  ...,
  legend_title = "grouping",
  to_return = "plot",
  translate = FALSE,
  sd_ratio = NULL,
  add_null_line = FALSE
)
}
\arguments{
\item{object}{model object}

\item{newdata}{data set to use for prediction (must include all predictors used in model object)}

\item{x_axis}{x-axis for ggplot}

\item{y_axis}{y-axis for ggplot, predicted value by default}

\item{grouping}{grouping variable for colour and shape of ggplot points (and optionally LOESS curve)}

\item{x_range}{if choosing "stata" for which_smooth, sequance of x-values for which to produce summary Stata-like margins}

\item{which_smooth}{optional parameter to add LOESS curve
* "both" = default; adds a LOESS overall and another LOESS grouped by grouping variable
* "single" = adds LOESS curve, ungrouped
* "by_grouping" = adds LOESS curve, grouped by the grouping variable
* any other input will not add a LOESS curve
* "stata" will produce Stata-like margins, using the calculate_stats() function}

\item{point_shape}{optional, point shape if choosing Stata-like margins. Default value is 16.}

\item{...}{optional, to pass on to the labs() in ggplot}

\item{legend_title}{Title for your legend, by default the grouping variable name}

\item{to_return}{optionally, tell make_margins() what to return
* "plot" is the default, and returns a ggplot object
* "df" returns the parameter newdata with the addition of a "predicted" column
* anything else will return only the "predicted column, which can be easily joined to newdata}

\item{translate}{TRUE/FALSE (FALSE is default). Should the clock residuals be translated back to unstandardised values?}

\item{sd_ratio}{optional; standard deviation ratio to translate standardised to unstandardised}

\item{add_null_line}{optional; adds a dashed red line at y = 0 if TRUE}
}
\value{
either the predicted values (with or without newdata) or a plot, as specified using the to_return argument
}
\description{
Produce predictive margins given a model object, new data, and, optionally, plot them.
For plotting, you must specify the x, y, and grouping variables.
Optionally, you can pass variables (...) into the labs() in ggplot.
}
\seealso{
[ggplot2::labs()] for more detailed instructions

[base::cbind()] for more detailed instructions
}
