% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{po.kable}
\alias{po.kable}
\title{Print out the coefficient comparison table}
\usage{
po.kable(tibble, caption = " ")
}
\value{
Returns a kable
}
\description{
print out the coefficient comparison table
}
