% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{return.fm}
\alias{return.fm}
\title{Compare linear regression models using ANOVA}
\usage{
return.fm(model1, model2, model3 = NULL)
}
\arguments{
\item{model1}{linear regression model object; full model}

\item{model2}{linear regression model object; reduced model}

\item{model3}{optional; linear regression model object; most reduced model}
}
\value{
Returns optimal linear regression model object.
}
\description{
Allows to compare up to three linear regression models. Returns the reduced model if ANOVA is statistically significant or the full model. Takes up to three models as input
}
