% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tabble}
\alias{tabble}
\title{Put dataframes together into an excel file}
\usage{
tabble(file.list, file.path, path.to.write.to, file.name)
}
\arguments{
\item{file.list}{list of .csv files (no path, just file names)}

\item{file.path}{path to the folder containing the list of .csv files to be combined}

\item{path.to.write.to}{path of the folder to which you want to write out the final .xlsx file}

\item{file.name}{name of .xlsx file to be made and writen out}
}
\value{
Returns a .xlsx file
}
\description{
put a list of .csv files into a .xlsx files, where each tab (sheet) is named after the csv
}
